@@ -6,35 +6,39 @@
     "babel-polyfill": "^6.3.14",
     "isomorphic-fetch": "^2.1.1",
     "lodash": "^4.0.0",
-    "react": "^0.14.7",
-    "react-addons-transition-group": "^0.14.6",
-    "react-dom": "^0.14.7",
+    "react": "^15.1.0",
+    "react-addons-transition-group": "^15.1.0",
+    "react-dom": "^15.1.0",
     "react-redux": "^4.2.1",
-    "react-router": "2.0.0",
+    "react-router": "^2.4.1",
     "react-router-redux": "^4.0.0",
     "redux": "^3.2.1",
     "redux-logger": "^2.6.1",
-    "redux-saga": "^0.9.3",
+    "redux-saga": "^0.10.5",
     "semantic-ui-less": "^2.1.7",
     "velocity-animate": "^1.2.3",
     "velocity-react": "^1.1.1"
   },
   "devDependencies": {
     "babel-core": "^6.6.0",
-    "babel-eslint": "^5.0.0",
+    "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
     "babel-polyfill": "^6.6.1",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.3.13",
     "babel-preset-react-hmre": "^1.0.1",
-    "eslint-plugin-react": "^4.1.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "eslint": "^2.11.1",
+    "eslint-plugin-react": "^5.1.1",
     "express": "^4.13.3",
+    "npm-install-webpack-plugin": "^4.0.1",
     "redux-devtools": "^3.1.0",
     "redux-devtools-dock-monitor": "^1.0.1",
     "redux-devtools-log-monitor": "^1.0.5",
     "webpack": "^1.9.11",
     "webpack-dev-middleware": "^1.2.0",
-    "webpack-hot-middleware": "^2.8.1"
+    "webpack-hot-middleware": "^2.8.1",
+    "webpack-notifier": "^1.3.0"
   },
   "scripts": {
     "start": "node server.js",
