@@ -1,5 +1,28 @@
-if (process.env.NODE_ENV === 'production') {
-	module.exports = require('./App.prod');
-} else {
-	module.exports = require('./App.dev');
+import React, { Component, PropTypes } from 'react';
+import { connect } from 'react-redux';
+import { browserHistory } from 'react-router';
+
+export class App extends Component {
+  render() {
+    const { children } = this.props;
+    return (
+      <div className="wrapper">
+        OK
+        {children}
+      </div>
+    );
+  }
 }
+
+App.propTypes = {
+  // Injected by React Router
+  children: PropTypes.node
+}
+
+function mapStateToProps(state, ownProps) {
+  return ownProps;
+}
+
+export default connect(mapStateToProps, {
+  //
+})(App);
