@@ -1,20 +1,16 @@
-import React { Component, PropTypes } from 'react';
+import React, { Component, PropTypes } from 'react';
 import { Provider } from 'react-redux';
 import routes from '../routes';
-import DevTools from './DevTools';
 import { Router } from 'react-router';
 
-export class App extends Component {
+export default class Root extends Component {
   render() {
     const { store, history } = this.props;
     return (
       <Provider store={store}>
-        <div>
-          <Router history={history} routes={routes} />
-          <DevTools />
-        </div>
+        <Router history={history} routes={routes}></Router>
       </Provider>
-    );
+    )
   }
 }
 
