@@ -1,15 +1,15 @@
 import 'babel-polyfill';
 import React from 'react';
-import { render } from 'react';
+import { render } from 'react-dom';
 import { browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
-import App from '.containters/App';
+import Root from './containers/Root';
 import configureStore from './store/configureStore';
 
-const store = configureStore();
+const store = configureStore(browserHistory);
 const history = syncHistoryWithStore(browserHistory, store);
 
 render(
-  <App store={store} history={history} />,
-  document.getElementById('app')
+  <Root store={store} history={history} />,
+  document.getElementById('root')
 );
